refactor: Отчеты
- добавил репозиторий
- немного подрефакторил код
