+++ app/Providers/AppServiceProvider.php
@@ -24,6 +24,9 @@ class AppServiceProvider extends ServiceProvider
      */
     public function register()
     {
-        //
+        $this->app->bind(
+            \App\Contracts\ReportRepositoryContract::class,
+            \App\Repositories\ReportRepositoryEloquent::class
+        );
     }
 }
